moved page, controller and router into wip folders
