+++ server/controllers/wip/genFields/genFields.ts
@@ -1,7 +1,7 @@
 import { Response } from 'express'
 import FormWizard from 'hmpo-form-wizard'
-import backUrl from '../../utils/backUrl'
-import FormInitialStep from '../base/formInitialStep'
+import backUrl from '../../../utils/backUrl'
+import FormInitialStep from '../../base/formInitialStep'
 
 export default class GenFields extends FormInitialStep {
   middlewareSetup() {
