+++ server/routes/index.ts
@@ -12,7 +12,7 @@ import genericRouter from './generic'
 import prisonerSearchRoutes from '../controllers/addPrisoner/prisonerSearch'
 import addPrisonerRouter from './addPrisoner'
 import questionsRouter from './questions/router'
-import genFieldsRouter from './genFields'
+import genFieldsRouter from './wip/genFields'
 
 export default function routes(services: Services): Router {
   const get = (path: string | string[], handler: RequestHandler) => router.get(path, asyncMiddleware(handler))
